oak: Refactor RequestStub to its own private header file
